+++ client-sdk/src/client/misc/payment_memo.rs
@@ -1,8 +1,5 @@
 use intmax2_interfaces::{
-    api::{
-        block_builder::interface::Fee, store_vault_server::interface::StoreVaultClientInterface,
-    },
-    data::encryption::Encryption,
+    api::store_vault_server::interface::StoreVaultClientInterface, data::encryption::Encryption,
 };
 
 use intmax2_zkp::common::{signature::key_set::KeySet, transfer::Transfer};
@@ -11,10 +8,6 @@ use serde::{de::DeserializeOwned, Deserialize, Serialize};
 use crate::client::error::ClientError;
 
 use super::get_topic;
-
-pub const WITHDRAWAL_FEE_MEMO: &str = "withdrawal_fee_memo";
-pub const CLAIM_FEE_MEMO: &str = "claim_fee_memo";
-pub const USED_OR_INVALID_MEMO: &str = "used_or_invalid_memo";
 
 #[derive(Default, Debug, Clone, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase", bound(deserialize = ""))]
@@ -25,25 +18,6 @@ pub struct PaymentMemo {
 }
 
 impl Encryption for PaymentMemo {}
-
-#[derive(Default, Debug, Clone, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct WithdrawalFeeMemo {
-    pub withdrawal_transfer: Transfer,
-    pub fee: Fee,
-}
-
-#[derive(Default, Debug, Clone, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct ClaimFeeMemo {
-    pub fee: Fee,
-}
-
-#[derive(Default, Debug, Clone, Serialize, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct UsedOrInvalidMemo {
-    pub reason: String,
-}
 
 pub async fn save_payment_memo<
     S: StoreVaultClientInterface,
