+++ client-sdk/src/client/mod.rs
@@ -2,6 +2,7 @@
 pub mod client;
 pub mod config;
 pub mod error;
+pub mod fee_payment;
 pub mod fee_proof;
 pub mod history;
 pub mod key_from_eth;
@@ -9,4 +10,3 @@ pub mod misc;
 pub mod receive_validation;
 pub mod strategy;
 pub mod sync;
-pub mod transfers_with_memo;
